requirements.txt
aiogram
aiohttp
python-dotenv
loguru
load-dotenv
pytz
gspread_asyncio
google-auth
asyncpg
colorlog
services
emoji
ruff